// frontend/src/types/Refaccion.ts
export interface Refaccion {
  id: string;
  nombre: string;
  codigo: string;
  categoria: string;
  estado: string;
  maquinas: string;
  medidas: string;
  cantidad: number;
  proveedor: string;
  ubicacion: string;
  historial: string;
  descripcion: string;
  imagen1?: string;
  imagen2?: string;
}